fix(script): PushInt normalises to int32 (TS toInt32 parity)
Players whose username yielded a composeUID with bit 31 set could only
land one hit on any NPC before the script's "Someone else is fighting
that." gate fired on every subsequent attack. About half of all
usernames produce such a UID, so this hit a large share of accounts.

Root cause: ScriptState.PushInt stored the raw Go int (e.g.
0x80000001 = 2147483649), but per-NPC varn storage is []int32. When
the player's melee/ranged attack script ran `~npc_retaliate(0)` and
that proc set `%npc_aggressive_player = uid`, the POP_VARN handler
truncated 2147483649 → int32(-2147483647). On the next tick's
player_in_combat_check the script pushed a fresh `uid` (still raw Go
int 2147483649) and compared it to the read-back varn value (sign-
extended back to Go int -2147483647). Mismatch → "Someone else is
fighting that.", same NPC, same player.

TS sidesteps this because every PushInt normalises through
toInt32 (`num | 0`, Numbers.ts:7), so both the fresh push and the
Int32Array round-trip share signed-int32 representation.

Fix: PushInt now stores `int(int32(v))`. Cheap, central, matches TS
exactly. RuneScript ints are 32-bit by design — wire protocol, varp,
and varn storage are all signed int32 — so normalising at the stack
boundary is the canonical layering.

Affects ALL UID values, ALL high-bit-set ints that round-trip through
varp/varn storage. Not just combat: any `=`/`\!`/`<`/`>` comparison
between a fresh PushInt and a previously-stored int32 value was
vulnerable.

One existing test (TestInvAdd_ObjStackValid_CountAboveStackLimit)
asserted the old Go-divergent error string "invalid count
(2147483648)"; updated to assert the post-cast "(-2147483648)". The
test still proves "count above StackLimit is rejected" — TS hits this
path the same way (toInt32 wraps 0x80000000 to -2147483648, falls into
the c<1 check rather than c>StackLimit). New
TestVarnRoundTripHighBitUidMatchesTS reproduces the user-visible bug
shape: OpUID → POP_VARN → PUSH_VARN → OpUID → assert top-two equal.

Third "comment claims TS-parity but doesn't" pattern this session
graduates into "Go diverges from TS in a way the comments never
acknowledged" — same lesson lineage as 30305239 (MetaScript), 35657661
(LoadSave), 6c7f7738 (effectiveApRange).

+++ pkg/script/state.go
@@ -438,13 +438,25 @@ type ScriptState struct {
 	SplitMesanim int32
 }
 
-// PushInt pushes v onto the int stack.
+// PushInt pushes v onto the int stack, normalised through signed-int32
+// (mirroring TS Numbers.ts:7 `toInt32(num) = num | 0`). RuneScript ints
+// are 32-bit; both varp/varn storage (`[]int32`) and the wire protocol
+// use signed-int32 representation, so any value pushed must share that
+// representation to round-trip through storage cleanly.
+//
+// Without this cast, high-bit-set values (e.g. `composeUID` results
+// >=0x80000000, which apply to ~50% of usernames) would compare unequal
+// after a write+read cycle: the fresh push stays as Go int 2147483649
+// while the varn read returns int(int32(stored)) = -2147483647. That
+// mismatch broke `%npc_aggressive_player == uid` in the combat-check
+// proc and surfaced as "Someone else is fighting that." after one hit.
+//
 // Panics if the stack is full (programming error / compiler bug).
 func (s *ScriptState) PushInt(v int) {
 	if s.ISP >= StackCapacity {
 		panic(fmt.Sprintf("script: int stack overflow at pc=%d in %q", s.PC, s.Script.Name))
 	}
-	s.IntStack[s.ISP] = v
+	s.IntStack[s.ISP] = int(int32(v))
 	s.ISP++
 }
